dashboard/requirements.txt
altair
pandas
streamlit

python-dotenv
redshift_connector